refactor(review): fix code-review findings on PR #729
Three findings from an xhigh code-review pass on the perf-audit PR:

1. .golangci.yml's gocritic hugeParam/rangeValCopy exclusion used
   text-pattern matching on the checker's own diagnostic string
   (fragile: silently stops matching if go-critic reformats its
   output). golangci-lint rejects combining disable-all with
   disabled-tags/disabled-checks, so switched to enabled-checks
   listing go-critic v0.14.3's full performance-tagged set minus
   those two checks — a structural exclusion keyed on checker
   identity instead of message text.

2. internal/index, internal/mdtext, and internal/schema each got a
   verbatim ~50-line structlayout_test.go (isPointerish +
   assertPointerFieldsFirst). Factored into a shared
   internal/structlayout package (mirroring the existing
   internal/testsymlink precedent: a non-_test.go file importing
   "testing", meant only for other packages' test files), with its
   own dedicated unit test. The three call sites now import it.

3. The new per-rule alloc-regression tests (maxfilelength,
   tokenbudget, orderedlistnumbering) lived in rule_test.go; moved
   each to a dedicated alloc_test.go, matching this repo's existing
   convention (internal/rules/tablereadability/alloc_test.go et al.).

go build/test/vet, golangci-lint, and mdsmith check . are all clean.
