+++ .golangci.yml
@@ -9,20 +9,6 @@ linters:
   exclusions:
     paths:
       - pkg/goldmark
-    rules:
-      # gocritic's hugeParam/rangeValCopy (both performance-tagged)
-      # flag ~260 existing by-value struct params and for-range loops
-      # across cmd/mdsmith, internal/build, internal/config, and
-      # internal/schema — genuine wins, but each fix touches a
-      # function signature or call sites, and some types are used
-      # across package boundaries. That is a much larger, separately
-      # reviewable pass than a single perf-audit PR; tracked as
-      # follow-up debt per docs/development/high-performance-go.md
-      # Tooling rather than grandfathered silently.
-      - linters: [gocritic]
-        text: "^hugeParam:"
-      - linters: [gocritic]
-        text: "^rangeValCopy:"
   enable:
     - errcheck
     - funlen
@@ -48,14 +34,39 @@ linters:
       min-complexity: 30
     gocritic:
       # docs/development/high-performance-go.md Tooling: "gocritic
-      # performance group are worth enabling". disable-all is
-      # required alongside enabled-tags: golangci-lint's gocritic
-      # wrapper treats enabled-tags as additive on top of the
-      # (much larger) style+diagnostic default set, not a
-      # restriction to just that tag.
+      # performance group are worth enabling". golangci-lint rejects
+      # combining disable-all with either disabled-tags or
+      # disabled-checks, so enabled-tags: [performance] can't be
+      # paired with a disabled-checks carve-out; enabled-checks
+      # spells out go-critic v0.14.3's full performance-tagged set
+      # (checkers/*.go Tags + checkers/rules/rules.go doc:tags)
+      # minus hugeParam/rangeValCopy instead.
+      #
+      # hugeParam/rangeValCopy flag 263 existing by-value struct
+      # params and for-range loops across cmd/mdsmith, internal/build,
+      # internal/config, and internal/schema — genuine wins, but each
+      # fix touches a function signature or call sites, and some
+      # types are used across package boundaries. That is a much
+      # larger, separately reviewable pass than a single perf-audit
+      # PR; tracked as follow-up debt per
+      # docs/development/high-performance-go.md Tooling rather than
+      # grandfathered silently. (golangci-lint's default
+      # max-same-issues/max-issues-per-linter caps hide the true
+      # count in a plain `run`; pass --max-issues-per-linter=0
+      # --max-same-issues=0 to see all of them.)
       disable-all: true
-      enabled-tags:
-        - performance
+      enabled-checks:
+        - appendCombine
+        - rangeExprCopy
+        - preferDecodeRune
+        - stringXbytes
+        - indexAlloc
+        - preferWriteByte
+        - preferFprint
+        - preferStringWriter
+        - sliceClear
+        - equalFold
+        - zeroByteRepeat
     perfsprint:
       # error-format (fmt.Errorf → errors.New) fires ~180 times
       # across the whole codebase, almost entirely on cold
