+++ internal/structlayout/structlayout.go
@@ -1,0 +1,65 @@
+// Package structlayout asserts that a struct type orders its
+// pointer-containing fields before its scalar fields. It is intended
+// for use only from *_test.go files.
+package structlayout
+
+import (
+	"reflect"
+	"testing"
+)
+
+// isPointerish reports whether t's GC ptrdata bitmap must mark this
+// field's word(s), directly or via a nested type.
+func isPointerish(t reflect.Type) bool {
+	switch t.Kind() {
+	case reflect.Ptr, reflect.String, reflect.Slice, reflect.Map,
+		reflect.Chan, reflect.Func, reflect.Interface, reflect.UnsafePointer:
+		return true
+	case reflect.Struct:
+		for i := 0; i < t.NumField(); i++ {
+			if isPointerish(t.Field(i).Type) {
+				return true
+			}
+		}
+		return false
+	case reflect.Array:
+		return t.Len() > 0 && isPointerish(t.Elem())
+	default:
+		return false
+	}
+}
+
+// firstOutOfOrderField returns the name of the first pointer-containing
+// field declared after a scalar-only field, and true. It returns ""
+// and false when every pointer-containing field precedes every
+// scalar-only field.
+func firstOutOfOrderField(typ reflect.Type) (string, bool) {
+	sawScalar := false
+	for i := 0; i < typ.NumField(); i++ {
+		f := typ.Field(i)
+		if isPointerish(f.Type) {
+			if sawScalar {
+				return f.Name, true
+			}
+			continue
+		}
+		sawScalar = true
+	}
+	return "", false
+}
+
+// AssertPointerFieldsFirst fails if a pointer-containing field is
+// declared after a scalar-only field. Go's GC ptrdata for a struct
+// spans from offset 0 through the end of the last pointer-containing
+// field; a scalar sandwiched before that point is scanned for
+// nothing, and a scalar declared after every pointer field costs
+// nothing. See docs/development/high-performance-go.md "Struct
+// layout".
+func AssertPointerFieldsFirst(t *testing.T, typ reflect.Type) {
+	t.Helper()
+	if field, bad := firstOutOfOrderField(typ); bad {
+		t.Fatalf("%s.%s (pointer-containing) is declared after a scalar field; "+
+			"reorder so all pointer fields precede scalar fields to minimize GC ptrdata",
+			typ.Name(), field)
+	}
+}
